+++ src/Core/Dfe.Complete.Application/Projects/Commands/TaskData/UpdateConfirmAcademyRiskProtectionArrangementsTaskCommand.cs
@@ -4,7 +4,7 @@ using Dfe.Complete.Application.Projects.Interfaces;
 using Dfe.Complete.Domain.Enums;
 using Dfe.Complete.Domain.Validators;
 using Dfe.Complete.Domain.ValueObjects;
-using Dfe.Complete.Utils;
+using Dfe.Complete.Utils.Exceptions;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 using System.ComponentModel.DataAnnotations;