+++ src/Frontend/Dfe.Complete/Pages/Projects/Decision/RecordDaoRevocation/Date/AddDaoRevocationDecisionDate.cshtml.cs
@@ -1,6 +1,6 @@
 using Dfe.Complete.Constants;
 using Dfe.Complete.Pages.Projects.Decision.RecordDaoRevocation.MinisterName;
-using Dfe.Complete.Services;
+using Dfe.Complete.Services.Interfaces;
 using GovUK.Dfe.CoreLibs.Caching.Interfaces;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
@@ -8,7 +8,7 @@ using System.ComponentModel.DataAnnotations;
 
 namespace Dfe.Complete.Pages.Projects.Decision.RecordDaoRevocation.Date
 {
-    public class AddDaoRevocationDecisionDateModel(ISender sender, ILogger<AddDaoRevocationMinisterNameModel> logger, ErrorService errorService,
+    public class AddDaoRevocationDecisionDateModel(ISender sender, ILogger<AddDaoRevocationMinisterNameModel> logger, IErrorService errorService,
         ICacheService<IMemoryCacheType> cacheService) : DaoRevocationProjectLayoutModel(sender, logger, cacheService)
     {
         [BindProperty]