+++ src/Frontend/Dfe.Complete/Pages/Projects/Transfer/CreateNewProject.cshtml.cs
@@ -1,20 +1,20 @@
+using Dfe.Complete.Application.Projects.Commands.CreateProject;
+using Dfe.Complete.Domain.Constants;
+using Dfe.Complete.Domain.ValueObjects;
+using Dfe.Complete.Extensions;
+using Dfe.Complete.Services.Interfaces;
+using Dfe.Complete.Validators;
+using MediatR;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
-using Dfe.Complete.Validators;
-using Dfe.Complete.Services;
 using System.ComponentModel.DataAnnotations;
-using Dfe.Complete.Application.Projects.Commands.CreateProject;
-using MediatR;
-using Dfe.Complete.Domain.ValueObjects;
-using Microsoft.AspNetCore.Authorization;
-using Dfe.Complete.Extensions;
-using NotFoundException = Dfe.Complete.Utils.NotFoundException;
-using Dfe.Complete.Domain.Constants;
+using NotFoundException = Dfe.Complete.Utils.Exceptions.NotFoundException;
 
 namespace Dfe.Complete.Pages.Projects.Transfer
 {
     [Authorize(policy: UserPolicyConstants.CanCreateProjects)]
-    public class CreateNewProjectModel(ISender sender, ErrorService errorService, ILogger<CreateNewProjectModel> logger)
+    public class CreateNewProjectModel(ISender sender, IErrorService errorService, ILogger<CreateNewProjectModel> logger)
         : PageModel
     {
         [BindProperty]